Switch to controller specific JavaScript
It turns out packing all the js files into single bundle is pure
evil when it is not a SPA. For instance, when hooking to the page
load event handlers are not bound since there are not such elements
on the main page. Decisions like using 3rd party js library (such
as arrive.js) led to necessity to bound to at least 2 events: page
load (as there could target page refreshes) and target element
arrival, which doesn't seem right. For the same reason turbolinks
has been also removed.