Rename Address.address column to street_address
Though naming Address.address is not critical when working with Active
Record, in particular, Address.new(address: '111 Smth',...) is
absolutely legal syntax. But, when it comes to Rectify Form Objects, its
from_params method stumbles when it encounters the enclosed 'address' tag
as the behaviour of this method is intented to be recursive. So it
'thinks' that under this 'address' lies the target hash to populate the
fields.